Break remaining runtime circular imports
The first TDZ fix (bT) addressed AppManager <-> DesktopManager only. A second
TDZ (sT) surfaced from four more runtime value<->value cycles:

- MoviePlayerContext & PictureViewerContext called classicyAppEventHandler
  back into AppManager (which itself imports both contexts)
- DesktopIconContext called classicyDesktopStateEventReducer and deFocusApps
  from AppManager (same cycle)
- SoundManagerUtils and ClassicySounds both imported each other as values
  (SoundDataSchema / ClassicySounds), and ClassicySounds calls
  SoundDataSchema.parse at module load
- DateAndTimeManagerUtils imports useAppManager/Dispatch from
  AppManagerUtils, which imports DefaultAppManagerState and
  classicyDesktopStateEventReducer from AppManager, which imports the
  DateAndTime event handler from the same file

Extracted three pure, dependency-free modules so the runtime graph is
acyclic:

- ClassicyAppHelpers.ts: openApp, loadApp, closeApp, deFocusApps, focusApp,
  activateApp, getDefaultAppForFileType. AppManager still re-exports them
  for its public API.
- ClassicySoundSchema.ts: SoundDataSchema + SoundData type.
- ClassicyDateAndTimeEventHandler.ts: classicyDateTimeManagerEventHandler.
  The hook-bearing DateAndTimeManagerUtils re-exports it for backwards
  compatibility.

Context files now import values from the pure helpers files and import
types from AppManager as type-only, so no runtime cycle can form.

All 451 tests pass.
