+++ src/SystemFolder/ControlPanels/SoundManager/ClassicySoundSchema.ts
@@ -1,0 +1,14 @@
+import { z } from "zod";
+
+export const SoundDataSchema = z.object({
+	src: z.array(z.string()),
+	sprite: z.record(
+		z.string(),
+		z.union([
+			z.tuple([z.number(), z.number()]),
+			z.tuple([z.number(), z.number(), z.boolean()]),
+		]),
+	),
+});
+
+export type SoundData = z.infer<typeof SoundDataSchema>;
