+++ src/index.ts
@@ -1,4 +1,5 @@
 export * from "./SystemFolder/ControlPanels/ClassicyControlPanels";
+export * from "./SystemFolder/ControlPanels/AppManager/ClassicyAppHelpers";
 export * from "./SystemFolder/ControlPanels/AppManager/ClassicyAppManager";
 export * from "./SystemFolder/ControlPanels/AppManager/ClassicyAppManagerContext";
 export * from "./SystemFolder/ControlPanels/AppManager/ClassicyAppManagerUtils";
@@ -12,12 +13,14 @@ export * from "./SystemFolder/ControlPanels/AppearanceManager/ClassicyIcons";
 export * from "./SystemFolder/ControlPanels/AppearanceManager/ClassicySize";
 export * from "./SystemFolder/ControlPanels/AppearanceManager/ClassicySounds";
 export * from "./SystemFolder/ControlPanels/AppearanceManager/ClassicyWallpapers";
+export * from "./SystemFolder/ControlPanels/DateAndTimeManager/ClassicyDateAndTimeEventHandler";
 export * from "./SystemFolder/ControlPanels/DateAndTimeManager/ClassicyDateAndTimeManager";
 export * from "./SystemFolder/ControlPanels/DateAndTimeManager/ClassicyDateAndTimeManagerUtils";
 export * from "./SystemFolder/ControlPanels/SoundManager/ClassicySoundManager";
 export * from "./SystemFolder/ControlPanels/SoundManager/ClassicySoundManagerContext";
 export * from "./SystemFolder/ControlPanels/SoundManager/ClassicySoundManagerProvider";
 export * from "./SystemFolder/ControlPanels/SoundManager/ClassicySoundManagerUtils";
+export * from "./SystemFolder/ControlPanels/SoundManager/ClassicySoundSchema";
 export * from "./SystemFolder/Finder/Finder";
 export * from "./SystemFolder/Finder/FinderContext";
 export * from "./SystemFolder/QuickTime/MoviePlayer/MoviePlayer";
